fix stats
