--- task-03/src/routes/stats.ts
@@ -1,7 +1,0 @@
-import express from "express";
-import controller from "../controllers/stats";
-const router = express.Router();
-
-router.get("/stats", controller.getStats);
-
-export = router;
